Generation for multiple players done and some miscellanous changes
The map generation is the same for every players, changed Lobby scene : created 4 empty gameobjects child of the player spawner, player spawner script changed to spawn players corresponding to the position of the child.
Add a ToUpper to the join code in the connection button script, the code is not case sensitive anymore.