Added basics wall textures and ground for the Rooms
Textures applied via materials in the room prefab.
Material for wall created as well as for ground.
Merged all ground piece of texture in one.
Changed working system for open and locked entries :
a plane in front of the entry wall is placed when locked, which is the size of a wall, it desactivates when the entry is open.